Updates to Game Integrations - Mike Tysons PunchOut
The Mike Tysons Punchout integration is evolving, and getting more complex. The current integration seems to cover the game well. Using it, and python NEAT The first 3 characters on MTPO were defeated by the AI so far. More tweaks are needed to proceed. Some example videos of the AI knocking out several characters are included.